Completed getElementsByClassName.js
It was a bit of a crude solution. I'm not 100% thrilled with it since I had to run checkName on document.body instead of letting checkNodes encompass that first check, but I did not want to add an if statement to check for it within checkNodes since it would force the interpreter to check for that condition each time when it was not necessary past the first.